% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLoanOrders.R
\name{getLoanOrders}
\alias{getLoanOrders}
\title{Get lending data}
\usage{
getLoanOrders(.symbol = "BTC", .limit = 10000)
}
\arguments{
\item{.symbol}{Symbol}

\item{.limit}{Limit}
}
\description{
This function returns the loan orders for a given symbol in the form of a data.frame
}
\examples{
getLoanOrders()
}
\keyword{LoanOrders}
