% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOHLCV.R
\name{getOHLCV}
\alias{getOHLCV}
\title{Get OHLCV data}
\usage{
getOHLCV(.symbol1 = "BTC", .symbol2 = "ETH", .period = 1800,
  .start = 1483228800, .end = 9999999999)
}
\arguments{
\item{.symbol1}{Symbol 1}

\item{.symbol2}{Symbol 2}

\item{.period}{Time-series data interval in seconds; default 30 minutes}

\item{.start}{Unix start time to be passed into API call; default 1483228800}

\item{.end}{Unix end time; default 9999999999}
}
\description{
This function returns Open-High-Low-Close-Volume data for a given pair.
}
\examples{
getOHLCV()
}
\keyword{OHLCV}
