% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOrderBook.R
\name{getOrderBook}
\alias{getOrderBook}
\title{Get order book}
\usage{
getOrderBook(.symbol1 = "BTC", .symbol2 = "ETH", .depth = 5000)
}
\arguments{
\item{.symbol1}{Symbol 1}

\item{.symbol2}{Symbol 2}

\item{.depth}{Depth}
}
\description{
This function returns the order book for a given pair in the form of a data.frame
}
\examples{
getOrderBook()
}
\keyword{OrderBook}
