% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoloniexPublic.R
\docType{methods}
\name{GetPoloniexPublicCommands,PoloniexPublicAPI-method}
\alias{GetPoloniexPublicCommands,PoloniexPublicAPI-method}
\title{Returns REST service commands for the Poloniex Public API}
\usage{
\S4method{GetPoloniexPublicCommands}{PoloniexPublicAPI}(theObject)
}
\arguments{
\item{theObject}{The public client API object on which the function should be called}
}
\value{
commands - a list with Key/value string pairs
 Represents the command strings for the respective public API functions.
}
\description{
Returns REST service commands for the Poloniex Public API
}
\examples{
poloniex.public <- PoloniexPublicAPI()
GetPoloniexPublicCommands(poloniex.public)
}
