% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoloniexPublic.R
\docType{methods}
\name{GetPoloniexPublicURL,PoloniexPublicAPI-method}
\alias{GetPoloniexPublicURL,PoloniexPublicAPI-method}
\title{Returns REST service URL for the Poloniex Public API}
\usage{
\S4method{GetPoloniexPublicURL}{PoloniexPublicAPI}(theObject)
}
\arguments{
\item{theObject}{The object on which the function should be called}
}
\value{
base.url - a length-one character vector.
 Represents the Poloniex public API service base url.
}
\description{
Returns REST service URL for the Poloniex Public API
}
\examples{
poloniex.public <- PoloniexPublicAPI()
GetPoloniexPublicURL(poloniex.public)
}
