% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoloniexPublic.R
\name{GetPoloniexPublicURL}
\alias{GetPoloniexPublicURL}
\title{Returns REST service URL for the Poloniex Public API}
\usage{
GetPoloniexPublicURL(theObject)
}
\arguments{
\item{theObject}{The object on which the function should be called}
}
\value{
base.url - a length-one character vector.
 Represents the Poloniex public API service base url.
}
\description{
Returns REST service URL for the Poloniex Public API
}
\examples{
poloniex.public <- PoloniexPublicAPI()
GetPoloniexPublicURL(poloniex.public)
}
