% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoloniexTrading.R
\docType{methods}
\name{GetPoloniexTradingCommands,PoloniexTradingAPI-method}
\alias{GetPoloniexTradingCommands,PoloniexTradingAPI-method}
\title{Returns REST service commands for the Poloniex Trading API}
\usage{
\S4method{GetPoloniexTradingCommands}{PoloniexTradingAPI}(theObject)
}
\arguments{
\item{theObject}{The Trading client API object on which the function should be called}
}
\value{
commands - a list with Key/value string pairs
 Represents the command strings for the respective Trading API functions.
}
\description{
Returns REST service commands for the Poloniex Trading API
}
\examples{
\dontrun{
poloniex.trading <- PoloniexTradingAPI(your.key, your.secret)
GetPoloniexTradingCommands(poloniex.trading)
}
}
