% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoloniexTrading.R
\name{GetPoloniexTradingCommands}
\alias{GetPoloniexTradingCommands}
\title{Returns REST service commands for the Poloniex Trading API}
\usage{
GetPoloniexTradingCommands(theObject)
}
\arguments{
\item{theObject}{The Trading client API object on which the function should be called}
}
\value{
commands - a list with Key/value string pairs
 Represents the command strings for the respective Trading API functions.
}
\description{
Returns REST service commands for the Poloniex Trading API
}
\examples{
poloniex.trading <- PoloniexTradingAPI()
GetPoloniexTradingCommands(poloniex.trading)
}
