% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoloniexTrading.R
\name{GetPoloniexTradingURL}
\alias{GetPoloniexTradingURL}
\title{Returns REST service URL for the Poloniex Trading API}
\usage{
GetPoloniexTradingURL(theObject)
}
\arguments{
\item{theObject}{The object on which the function should be called}
}
\value{
trading.base.url - a length-one character vector.
 Represents the Poloniex Trading API service base url.
}
\description{
Returns REST service URL for the Poloniex Trading API
}
\examples{
poloniex.trading <- PoloniexTradingAPI()
GetPoloniexTradingURL(poloniex.trading)
}
