% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoloniexPublic.R
\docType{class}
\name{PoloniexPublicAPI-class}
\alias{PoloniexPublicAPI-class}
\alias{PoloniexPublicAPI}
\title{An S4 class representing the Poloniex Public API}
\description{
An S4 class representing the Poloniex Public API
}
\section{Slots}{

\describe{
\item{\code{base.url}}{A length-one numeric vector.
Represents the poloniex public API service base url.}

\item{\code{commands}}{A list containing Key/Value string pairs.
Represents the command strings of the associated public API functions.}
}}

