% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoloniexR.R
\docType{package}
\name{PoloniexR}
\alias{PoloniexR}
\alias{PoloniexR-package}
\title{R wrapper for the Poloniex Public and Trading API.}
\description{
This package provides a user-friendly R wrapper on top of the
Poloniex (Cryptocurrency Trading) REST API. The JSON results of the API calls are translated into R
Datastructures and are subsequently returned to the user in an intuitive manner.
}
\details{
The package provides the user with two separate classes:
 - PoloniexPublicAPI: Allows the user to call Poloniex Public API functions.
 - PoloniexTradingAPI: Allows the user to call Poloniex trading API functions.

Examples:
 https://github.com/VermeirJellen/PoloniexR
 http://EssentialQuant.com/blog/PoloniexR
}
