% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoloniexTrading.R
\docType{class}
\name{PoloniexTradingAPI-class}
\alias{PoloniexTradingAPI-class}
\alias{PoloniexTradingAPI}
\title{An S4 class representing the Poloniex Trading API.}
\description{
All calls to the trading API are sent via HTTP POST.
The POST data itself is signed with your key's secret
according to the HMAC-SHA512 method.
}
\section{Slots}{

\describe{
\item{\code{trading.base.url}}{A length-one character vector.
Represents the poloniex trading API web service base url.}

\item{\code{key}}{A length-one character vector.
Represents the public key of your Poloniex trading account}

\item{\code{secret}}{A length-one character vector.
Represents the private key of your Poloniex trading account}

\item{\code{commands}}{A list containing Key/Value string pairs.
Represents the command strings of the associated trading API functions.}
}}

