% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoloniexTrading.R
\docType{methods}
\name{ReturnBalances,PoloniexTradingAPI-method}
\alias{ReturnBalances,PoloniexTradingAPI-method}
\title{Returns all the available non-zero balances associated with
this account.}
\usage{
\S4method{ReturnBalances}{PoloniexTradingAPI}(theObject)
}
\arguments{
\item{theObject}{The Trading client API object on which the function should be called.}
}
\value{
a vector containing non-zero currency balances. Each
 entry corresponds to a balance for a single currency.
}
\description{
Returns all the available non-zero balances associated with
this account.
}
\examples{
\dontrun{
poloniex.trading <- PoloniexTradingAPI(your.key, your.secret)
balances         <- ReturnBalances(poloniex.trading)
head(balances)
}
}
