% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoloniexTrading.R
\name{ReturnBalances}
\alias{ReturnBalances}
\title{Returns all the available non-zero balances associated with
this account.}
\usage{
ReturnBalances(theObject)
}
\arguments{
\item{theObject}{The Trading client API object on which the function should be called.}
}
\value{
a vector containing non-zero currency balances. Each
 entry corresponds to a balance for a single currency.
}
\description{
Returns all the available non-zero balances associated with
this account.
}
\examples{
poloniex.trading <- PoloniexTradingAPI()
balances         <- ReturnBalances(poloniex.trading)
head(balances)
}
