% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoloniexTrading.R
\name{ReturnCompleteBalances}
\alias{ReturnCompleteBalances}
\title{Returns all of your account balances, including available balance, balance on orders,
and the estimated BTC value of your balance.
By default, this call is limited to your exchange account; set the all.balances input parameter to TRUE
to include your margin and lending accounts.}
\usage{
ReturnCompleteBalances(theObject, all.balances = FALSE)
}
\arguments{
\item{theObject}{The Trading client API object on which the function should be called.}

\item{all.balances}{logical - set to TRUE to also return margin and lending
account balances.}
}
\value{
todo
}
\description{
Returns all of your account balances, including available balance, balance on orders,
and the estimated BTC value of your balance.
By default, this call is limited to your exchange account; set the all.balances input parameter to TRUE
to include your margin and lending accounts.
}
\examples{
poloniex.trading <- PoloniexTradingAPI()
balances         <- ReturnCompleteBalances(poloniex.trading)
head(balances)

balances.all <- ReturnCompleteBalances(poloniex.trading)
head(balances.all)
}
