% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoloniexPublic.R
\name{ReturnCurrencies}
\alias{ReturnCurrencies}
\title{Returns information about currencies}
\usage{
ReturnCurrencies(theObject)
}
\arguments{
\item{theObject}{The object on which the function should be called}
}
\value{
A dataframe with currency information
 - rows: Information about a single currency
 - columns: id, txfee, minConf, disabled, delisted, frozen
}
\description{
Returns information about currencies
}
\examples{
poloniex.public <- PoloniexPublicAPI()
currencies <- ReturnCurrencies(poloniex.public)
}
