% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoloniexPublic.R
\docType{methods}
\name{ReturnTicker,PoloniexPublicAPI-method}
\alias{ReturnTicker,PoloniexPublicAPI-method}
\title{Returns the ticker for all markets}
\usage{
\S4method{ReturnTicker}{PoloniexPublicAPI}(theObject)
}
\arguments{
\item{theObject}{The public client API object on which the function should be called}
}
\value{
A dataframe containing ticker information
 Rows    - Cryptocurrency Pairs
 Columns - id, last, lowestAsk, highestBid, percentChange
  baseVolume, quoteVolume, isFrozen, high24hr, low24hr
}
\description{
Returns the ticker for all markets
}
\examples{
poloniex.public <- PoloniexPublicAPI()
GetPoloniexPublicURL(poloniex.public)
}
