% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoloniexPublic.R
\docType{methods}
\name{SetPoloniexPublicCommands,PoloniexPublicAPI-method}
\alias{SetPoloniexPublicCommands,PoloniexPublicAPI-method}
\title{Setter for REST service commands for the Poloniex public API.}
\usage{
\S4method{SetPoloniexPublicCommands}{PoloniexPublicAPI}(theObject, commands)
}
\arguments{
\item{theObject}{The public client API object on which the function should be called}

\item{commands}{-  a list with Key/value string pairs
Represents the command strings for the respective public API functions.}
}
\value{
theObject - a copy of the object with the modified command slot
}
\description{
Setter for REST service commands for the Poloniex public API.
}
\examples{
poloniex.public <- PoloniexPublicAPI()
GetPoloniexPublicCommands(poloniex.public)
commands.new <- list(returnTicker = "returnTicker.new",
                    return24hVolume = "return24hVolume.new")
poloniex.public <- SetPoloniexPublicCommands(poloniex.public,
                                             commands = commands.new)
GetPoloniexPublicCommands(poloniex.public)
}
